Move OllamaServerInfos class to base module
- Eliminate dependency of the core module on the API module.
